% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/its_analysis.R
\name{add_proportion_columns}
\alias{add_proportion_columns}
\title{add_proportion_columns}
\usage{
add_proportion_columns(agg_df)
}
\arguments{
\item{agg_df}{the wide-aggregated data for its analysis}
}
\value{
agg_df with variables expressed as proportions of count
}
\description{
add_proportion_columns
}
