% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning.R
\name{age_band_recode}
\alias{age_band_recode}
\title{age_band_recode}
\usage{
age_band_recode(df)
}
\arguments{
\item{df}{dataframe containing AgeBand column to be recoded}
}
\value{
df with AgeBand recoded
}
\description{
age_band_recode
}

