% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bundle_data_cleaning.R
\name{bundle_element_data_completeness}
\alias{bundle_element_data_completeness}
\title{bundle_element_data_completeness}
\usage{
bundle_element_data_completeness(bundles = clahrcnwlhf::bundle_data_clean)
}
\arguments{
\item{bundles}{the bundle data to be assessed for completeness}
}
\value{
bundles with additional column showing number of NAs
}
\description{
bundle_element_data_completeness
}
