% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_conversion.R
\name{check_idempotency_of_convert_time_cols}
\alias{check_idempotency_of_convert_time_cols}
\title{check_idempotency_of_convert_time_cols}
\usage{
check_idempotency_of_convert_time_cols(df, lt_to_lt = TRUE)
}
\arguments{
\item{df}{dataframe}

\item{lt_to_lt}{specify starting point of conversion}
}
\value{
logical indicating whether idempotent
}
\description{
check_idempotency_of_convert_time_cols
}
