% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning.R
\name{clean_and_save}
\alias{clean_and_save}
\title{clean_and_save}
\usage{
clean_and_save()
}
\description{
Runs clean_data on admission_data and saves the result as
admission_data_clean, overwriting any previous version.
}

