% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bundle_data_cleaning.R
\name{clean_bundle_data}
\alias{clean_bundle_data}
\title{clean_bundle_data}
\usage{
clean_bundle_data(df)
}
\arguments{
\item{df}{the raw bundle data}
}
\value{
the cleaned bundle data
}
\description{
clean_bundle_data
}
