% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning.R
\name{clean_data}
\alias{clean_data}
\title{clean_data}
\usage{
clean_data(df, restrict_disch_date = TRUE)
}
\arguments{
\item{df}{the data frame containing the episode data to be cleaned}
}
\value{
A data frame containing the cleaned data
}
\description{
This function cleans the raw data by performing the following steps:
\enumerate{
  \item Remove leading and trailing whitespace from any character fields!
  \item Replace any instances of NULL or #N/A in character fields with NA
  \item Convert the following columns to factors: AgeBand, EthnicGroup, Sex,
  AdmissionMethodCode, AdmissionType, CSPLastWard
  \item Convert the following columns from character to Date format:
  the following columns of the raw data:
  AdmissionDate, DischargeDate, EpisodeStartDate, EpisodeEndDate}
}
