% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nicor_data_cleaning.R
\name{clean_nicor_data}
\alias{clean_nicor_data}
\title{clean_nicor_data}
\usage{
clean_nicor_data(df)
}
\arguments{
\item{df}{the raw nicor data}
}
\value{
the cleaned nicor data
}
\description{
clean_nicor_data
}

