% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/its_analysis.R
\name{column_as_proportion}
\alias{column_as_proportion}
\title{column_as_proportion}
\usage{
column_as_proportion(df, numerator_col, denominator_col, prop_col_name)
}
\arguments{
\item{df}{input dataframe}

\item{numerator_col}{column of df to form numerator of new proportion column}

\item{denominator_col}{column of df to form denominator of new proportion column}

\item{prop_col_name}{name for new proportion column}
}
\value{
df with new proportion column added
}
\description{
column_as_proportion
}
