% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_analysis.R
\name{consecutive.diff}
\alias{consecutive.diff}
\title{consecutive diff}
\usage{
consecutive.diff(df, c1, c2, nm = "consec.diff")
}
\arguments{
\item{df}{the dataframe}

\item{c1}{a column name of a numerical column}

\item{c2}{a column name of a numerical column}

\item{nm}{new column name}
}
\value{
a dataframe which is a copy of df with a new column added, with name
given by the parameter nm. Each entry of this new column is equal to [the
entry in that row of c2] - [the entry in c1 from the row above]
}
\description{
consecutive diff
}

