% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_conversion.R
\name{convert_time_cols}
\alias{convert_time_cols}
\title{convert_time_cols}
\usage{
convert_time_cols(df, lt_to_ct = TRUE)
}
\arguments{
\item{df}{dataframe}

\item{lt_to_ct}{specify which way to convert}
}
\value{
df with times converted
}
\description{
convert_time_cols
}
