% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning.R
\name{convert_to_date}
\alias{convert_to_date}
\title{convert_to_date}
\usage{
convert_to_date(v, match_reg = "^[0-9]{4}-[0-9]{2}-[0-9]{2}",
  date_form = "\%Y-\%m-\%d")
}
\arguments{
\item{v}{character vector containing dates as substrings}

\item{match_reg}{regular expression specifying the format of the dates
contained within the substrings. Default is ^[0-9]{4}-[0-9]{2}-[0-9]{2}.}

\item{date_form}{the date format to be used for conversion.}
}
\value{
A date vector the same length as v
}
\description{
Convert a character vector containing dates as substrings into a vector of the
corresopding dates
}

