% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning.R
\name{convert_to_datetime}
\alias{convert_to_datetime}
\title{convert_to_datetime}
\usage{
convert_to_datetime(dv, tv, dt_form = "\%Y-\%m-\%d \%H\%M")
}
\arguments{
\item{dv}{a vector of dates or characters representing dates}

\item{tv}{a vector of times or characters representing times}

\item{dt_form}{the concatenated format of dv and tv}
}
\value{
A POSIXlt vector combining dv and tv.
}
\description{
convert_to_datetime
}
