% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/its_analysis.R
\name{create_readms}
\alias{create_readms}
\title{create_readms}
\usage{
create_readms(df)
}
\arguments{
\item{df}{dataframe of spell data}
}
\value{
df with readmission variable columns added
}
\description{
create_readms
}

