% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_analysis.R
\name{disch_time_table}
\alias{disch_time_table}
\title{disch_time_table}
\usage{
disch_time_table(df, split_by = "\%Y-\%m", discharge = TRUE)
}
\arguments{
\item{df}{dataframe in question}

\item{split_by}{the time period to split by. Defaults to months.}
}
\value{
a vector of row counts for each time period
}
\description{
Produces a table of number of rows of a dataframe df per time period of the
column DischargeDate
}
