% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/its_analysis.R
\name{drop_unnecessary_cols}
\alias{drop_unnecessary_cols}
\title{drop_unnecessary_cols}
\usage{
drop_unnecessary_cols(df, drop.cols = c("AdmissionDate", "DischargeDate",
  "PrimaryDiagnosis", "SecondaryDiagnosis1", "SecondaryDiagnosis2",
  "SecondaryDiagnosis3", "SecondaryDiagnosis4", "SecondaryDiagnosis5",
  "SecondaryDiagnosis6", "SecondaryDiagnosis7", "SecondaryDiagnosis8",
  "SecondaryDiagnosis9", "EthnicGroup", "EpisodeNumber", "EpisodeStartDate",
  "EpisodeStartTime", "EpisodeEndDate", "EpisodeEndTime", "extract",
  "NewPseudo2", "OldPseudoID", "new.pat", "new_spell", "episode_order"))
}
\arguments{
\item{df}{}
}
\value{
copy of df but with only those columns needed for ITS analysis
}
\description{
drop_unnecessary_cols
}
