% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bundle_linking.R
\name{dupe_link_details}
\alias{dupe_link_details}
\title{dupe_link_details}
\usage{
dupe_link_details(dupe_bundles)
}
\arguments{
\item{dupe_bundles}{output of duplicated_links}
}
\value{
dupe_bundles with details of the multiple bundles associated with each spell
}
\description{
dupe_link_details
}
