% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_analysis.R
\name{ethn_time_table}
\alias{ethn_time_table}
\title{ethn_time_table}
\usage{
<<<<<<< HEAD
ethn_time_table(df, split_by = "\%Y-\%m", ethn_col = "EthnicGroup")
}
\arguments{
\item{df}{a data frame of spells with column "EthnicGroup" or other specified
containing data on ethnic group for each spell}
=======
ethn_time_table(df, split_by = "\%Y-\%m")
}
\arguments{
\item{df}{a data frame of spells with column "EthnicGroup"}
>>>>>>> Began work on ethnicity analysis, adding necessary files first helper function.

\item{split_by}{date format to split by}
}
\value{
table of counts of ethnic groups by discharge time period
}
\description{
ethn_time_table
}

