% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_analysis.R
\name{group_by_date}
\alias{group_by_date}
\title{group_by_date}
\usage{
group_by_date(df, date_col = "DischargeDate", cutoff_dates, new_col,
  period_labels)
}
\arguments{
\item{df}{the dataframe}

\item{date_col}{a date column in df}

\item{cutoff_dates}{vector of dates}

\item{new_col}{name for new column}

\item{period_labels}{vector of character labels}
}
\value{
df with a new column added for the period date_col falls into for
that row. Periods are defined by cutoff_dates and labelled by period_labels.
}
\description{
group_by_date
}

