% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/its_analysis.R
\name{its_aggregate_wide}
\alias{its_aggregate_wide}
\title{its_aggregate_wide}
\usage{
its_aggregate_wide(df = clahrcnwlhf::emergency_adms, from.vignette = FALSE,
  ward_hf_types = NULL, ward_sites = NULL, imd_lookup = NULL)
}
\arguments{
\item{df}{emergency admissions episode data}
}
\value{
aggregated data for ITS analysis
}
\description{
its_aggregate_wide
}
