% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_analysis.R
\name{lag.column}
\alias{lag.column}
\title{lag.column}
\usage{
lag.column(df, c1, nm = "lag.col", backlag = FALSE, sc = NULL)
}
\arguments{
\item{df}{a dataframe}

\item{c1}{name of a column of df}

\item{nm}{new column name}

\item{backlag}{direction of lagging}

\item{sc}{a screening column}
}
\value{
a copy of df with a new column in the dataframe that is equal to c1
but shifted either down (backlag = FALSE) or up (backlag = TRUE).
Furthermore, if a "screening" column sc is specified, then:
For backlag = FALSE, rows that are just above a TRUE in sc are not shifted
down - the row with the TRUE in has an NA in the new column (think new
patient rows when defining end of last episode)
For backlag = TRUE, rows that have a TRUE in sc are not shifted
up - the row above the TRUE has an NA in the new column (think new patient
rows when defining start of next episode)
}
\description{
lag.column
}

