% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bundle_linking.R
\name{link_and_save}
\alias{link_and_save}
\title{link_and_save}
\usage{
link_and_save(episodes = clahrcnwlhf::emergency_adms,
  bundles = clahrcnwlhf::bundle_data_clean,
  nicor = clahrcnwlhf::nicor_data_clean)
}
\arguments{
\item{episodes}{the emergency admissions data}

\item{bundles}{the care bundle data}

\item{nicor}{the nicor data}
}
\value{
Saves linked bundle and nicor data to the package
}
\description{
link_and_save
}
