% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bundle_linking.R
\name{link_bundles}
\alias{link_bundles}
\title{link_bundles}
\usage{
link_bundles(bundles = clahrcnwlhf::bundle_data_clean,
  episodes = clahrcnwlhf::emergency_adms, show.working = TRUE,
  bundle_date_col = "Admission.Datetime", bundle_id_col = "PseudoID",
  episode_id_col = "PseudoID", episode_date_col = "CSPAdmissionTime",
  episode_date_disch_col = "CSPDischargeTime",
  episode_spell_col = "spell_number", episode_new_spell = "new_spell")
}
\arguments{
\item{bundles}{the bundle data}

\item{episodes}{the episode data}
}
\value{
dataframe with new column linked.spell containing spell_number of the spell associated with this bundle
}
\description{
link_bundles
}
