% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bundle_linking.R
\name{link_diag_table}
\alias{link_diag_table}
\title{link_diag_table}
\usage{
link_diag_table(bis, lag_cutoff = 3)
}
\arguments{
\item{bis}{output of bundle_in_spell}

\item{lag_cutoff}{cutoff for classifying lags}
}
\value{
table of lag diagnostic information
}
\description{
link_diag_table
}

