% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bundle_linking.R
\name{link_nicor}
\alias{link_nicor}
\title{link_nicor}
\usage{
link_nicor(nicor = clahrcnwlhf::nicor_data_clean,
  episodes = clahrcnwlhf::emergency_adms, show.working = TRUE,
  bundle_date_col = "Date.of.Visit", bundle_id_col = "PseudoID",
  episode_id_col = "PseudoID", episode_date_col = "CSPAdmissionTime",
  episode_date_disch_col = "CSPDischargeTime",
  episode_spell_col = "spell_number", episode_new_spell = "new_spell")
}
\arguments{
\item{nicor}{the NICOR data}

\item{episodes}{dataframe of hospital admissions from data warehouse}

\item{show.working}{include helper columns in output}

\item{bundle_date_col}{column in NICOR data containing date pertaining to admission}

\item{bundle_id_col}{column in NICOR data containing patient (pseudo-)ID}

\item{episode_id_col}{column in episodes containing patient (pseudo-)ID}

\item{episode_date_col}{column in episodes containing datetime of admission}

\item{episode_date_disch_col}{column in episodes containing datetime of discharge}

\item{episode_spell_col}{column in episodes containing spell ID to return}

\item{episode_new_spell}{column in episodes indicating first episodes in spell}
}
\value{
nicor with linked spells, and helper columns if requested
}
\description{
link_nicor
}

