% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_analysis.R
\name{make.spellnumber.2}
\alias{make.spellnumber.2}
\title{make.spellnumber.2}
\usage{
make.spellnumber.2(df)
}
\arguments{
\item{df}{episode dataframe}
}
\value{
df with added column for spell number
}
\description{
make.spellnumber.2
}

