% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_analysis.R
\name{make_anydiag_flag}
\alias{make_anydiag_flag}
\title{make_anydiag_flag}
\usage{
make_anydiag_flag(dataframe, col_regex, new_colname, code_regex)
}
\arguments{
\item{dataframe}{the episode data}

\item{col_regex}{regex to match diagnosis column names}

\item{new_colname}{the name for the new flag column}

\item{code_regex}{the regex to match the diagnosis codes}
}
\value{
dataframe with a new column appended, indicating presence of a
diagnosis code matching code_regex in any of the columns whose names match
col_regex.
}
\description{
make_anydiag_flag
}
