% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_analysis.R
\name{make_diag_flag}
\alias{make_diag_flag}
\title{make_diag_flag}
\usage{
make_diag_flag(dataframe, colname = "PrimaryDiagnosis", code_regex,
  new_colname = "diag.flag")
}
\arguments{
\item{new_colname}{}
}
\value{
dataframe with diagnosis flag column added
}
\description{
make_diag_flag
}

