% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_analysis.R
\name{make_diag_flags}
\alias{make_diag_flags}
\title{make_diag_flags}
\usage{
make_diag_flags(dataframe, col_names, new_colname_suffix, code_regex)
}
\arguments{
\item{code_regex}{}
}
\description{
make_diag_flags
}

