% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_analysis.R
\name{make_emergency_adms_dataset}
\alias{make_emergency_adms_dataset}
\title{make_emergency_adms_dataset}
\usage{
make_emergency_adms_dataset(adc)
}
\arguments{
\item{adc}{the cleaned admissions data}
}
\value{
the emergency admissions dataset with additional columns for analysis
}
\description{
make_emergency_adms_dataset
}

