% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning.R
\name{make_factors}
\alias{make_factors}
\title{make_factors}
\usage{
make_factors(df, cols)
}
\arguments{
\item{df}{the dataframe whose columns are to be converted}

\item{cols}{a logical vector with one element for each column}
}
\value{
A dataframe identical to df except that any columns whose
corresponding entry in cols is \code{TRUE} is converted to a factor
}
\description{
Converts the specified columns of a dataframe to factors
}
