% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/its_analysis.R
\name{make_flag_variables}
\alias{make_flag_variables}
\title{make_flag_variables}
\usage{
make_flag_variables(df)
}
\arguments{
\item{df}{emergency admissions dataframe, with additional variables added by create_new_vars}
}
\value{
df with additional flag variables needed for the ITS analysis
}
\description{
make_flag_variables
}
