% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_analysis.R
\name{make_period_col}
\alias{make_period_col}
\title{make_period_col}
\usage{
make_period_col(df, colname, split_dates, period_labels, new_colname)
}
\arguments{
\item{df}{data frame to which we want to add a period column}

\item{colname}{column in df containing all the dates on which the period established}

\item{split_dates}{vector of strings representing the date cut-offs of the periods.
Note the first entry must a date before the earliest date in colname
and the last date must be a date after the final date in colname}

\item{period_labels}{labels for periods split by split_dates}

\item{new_colname}{name of new column created containing the period into which the date falls}
}
\value{
df with an extra column of periods
}
\description{
make_period_col
}

