% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning.R
\name{missing_data_table}
\alias{missing_data_table}
\title{missing_data_table}
\usage{
missing_data_table(df, split_by = "\%Y", result = "both")
}
\arguments{
\item{df}{the admissions data frame to be analysed for missing data}
}
\value{
a matrix showing missing data for each field by year
}
\description{
missing_data_table
}

