% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning.R
\name{na_count}
\alias{na_count}
\title{na_count}
\usage{
na_count(df)
}
\arguments{
\item{df}{the dataframe in question}
}
\value{
A dataframe showing the number of NA values in each column of df
}
\description{
Display the number of NAs per column
}
