% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_analysis.R
\name{new.pat}
\alias{new.pat}
\title{new.pat}
\usage{
new.pat(df, id = "PseudoID", adt = "CSPAdmissionTime",
  epno = "EpisodeNumber")
}
\arguments{
\item{df}{an episode dataframe}

\item{id}{the column name of the id field}

\item{epno}{the episode number in the spell}

\item{a.dt}{the column name of the admission datetime field}
}
\value{
a dataframe which is a copy of df with a new column
"new.pat" which is TRUE if the row represents an
episode that is the first for that PAT_CODE, and
FALSE if the row above represents an episode for
the same patient.
}
\description{
new.pat
}
