% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_analysis_3.R
\name{plot_age_dist}
\alias{plot_age_dist}
\title{plot_age_dist}
\usage{
plot_age_dist(df, stratify = FALSE)
}
\arguments{
\item{df}{data frame containing the data to be plotted}
}
\value{
plot of age distribution
}
\description{
plot_age_dist
}
