% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bundle_linking.R
\name{plot_lag_dist}
\alias{plot_lag_dist}
\title{plot_lag_dist}
\usage{
plot_lag_dist(bundles = clahrcnwlhf::bundle_data_clean,
  episodes = clahrcnwlhf::emergency_adms, bis = NULL, prev = TRUE,
  cumulative = FALSE)
}
\arguments{
\item{bundles}{dataframe of care bundle audit sheets}

\item{episodes}{dataframe of hospital admissions from data warehouse}

\item{bis}{optional pre-analysed output of bundle_in_spell}

\item{prev}{toggle between lags from previous admission and to next}

\item{cumulative}{show cumulative distributions vs distributions}
}
\value{
plot of the lag distribution
}
\description{
plot_lag_dist
}

