% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_analysis.R
\name{recode_ethnicity}
\alias{recode_ethnicity}
\title{recode_ethnicity}
\usage{
recode_ethnicity(df, ethn_col = "EthnicGroup")
}
\arguments{
\item{df}{a dataframe}

\item{ethn_col}{the column of df containing the ethnicity data to be recoded}
}
\value{
df with an additional column "EthnicGroupComp" holding the recoded ethnicity data
}
\description{
recode_ethnicity
}
