% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning.R
\name{remove_dupes}
\alias{remove_dupes}
\title{remove_dupes}
\usage{
remove_dupes(df, dupe_cols = c("PseudoID", "AdmissionDate",
  "CSPAdmissionTime", "DischargeDate", "CSPDischargeTime",
  "PrimaryDiagnosis", "SecondaryDiagnosis1", "SecondaryDiagnosis2",
  "EpisodeNumber"))
}
\arguments{
\item{df}{an episode dataframe to be de-duped}
}
\value{
the de-duped df
}
\description{
remove_dupes
}
