% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bundle_linking.R
\name{remove_duplicate_bundle_links}
\alias{remove_duplicate_bundle_links}
\title{remove_duplicate_bundle_links}
\usage{
remove_duplicate_bundle_links(linked_bundles)
}
\arguments{
\item{linked_bundles}{output of link_bundles}
}
\value{
linked_bundles with duplicate links removed
}
\description{
remove_duplicate_bundle_links
}

