% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning.R
\name{replace_ids}
\alias{replace_ids}
\title{replace_ids}
\usage{
replace_ids(df, id_group, id_col = "PseudoID")
}
\arguments{
\item{df}{dataframe of episode data}

\item{id_group}{group of ids to be merged}
}
\value{
df with PseudoIDs merged
}
\description{
replace_ids
}
