% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_analysis.R
\name{restrict_to_spells}
\alias{restrict_to_spells}
\title{restrict_to_spells}
\usage{
restrict_to_spells(df)
}
\arguments{
\item{df}{dataframe}
}
\value{
df restricted to only rows where 'new.spell' = TRUE
}
\description{
restrict_to_spells
}
