% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/its_analysis.R
\name{save_its_raw_data_csv}
\alias{save_its_raw_data_csv}
\title{save_its_raw_data_csv}
\usage{
save_its_raw_data_csv()
}
\value{
returns the raw data for ITS analysis invisibly
}
\description{
Saves a copy of the raw data for ITS analysis to the data-out folder.
This dataset is the output from create_new_vars(), filtered to
discharges prior to August 2017, with only columns required for
the ITS analysis
}
