% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning.R
\name{subset_by_date}
\alias{subset_by_date}
\title{subset_by_date}
\usage{
subset_by_date(df, col_name = "DischargeDate", start_date, end_date)
}
\arguments{
\item{df}{the dataframe to be subsetted}

\item{col_name}{the date column to be subsetted on}

\item{start_date}{the start date for the subsetting}

\item{end_date}{the end date for the subsetting}
}
\value{
a row subset of df, including all rows where col_name is between
start_date and end_date (inclusive).
}
\description{
subset_by_date
}

