% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_analysis_3.R
\name{test_plot_dist_table}
\alias{test_plot_dist_table}
\title{test_plot_dist_table}
\usage{
test_plot_dist_table(df, age_col = "AgeBand", stratify = FALSE,
  strat_col)
}
\arguments{
\item{df}{dataframe}

\item{age_col}{column containing age band variable}

\item{stratify}{indicate whether table is split into periods or overall}

\item{strat_col}{name of column to split age band by}
}
\value{
a (two-way) summary table of age band distribution containing frequncies for each period.date and total sums for both periods
}
\description{
test_plot_dist_table
}
