% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bundle_linking.R
\name{bundle_in_spell}
\alias{bundle_in_spell}
\title{bundle_in_spell}
\usage{
bundle_in_spell(bundles, episodes = clahrcnwlhf::emergency_adms,
  bundle_date_col = "Admission.Datetime", bundle_id_col = "PseudoID",
  episode_id_col = "PseudoID", episode_date_col = "CSPAdmissionTime",
  episode_date_disch_col = "CSPDischargeTime",
  episode_spell_col = "spell_number", episode_new_spell = "new_spell")
}
\arguments{
\item{bundles}{dataframe of care bundle audit sheets}

\item{episodes}{dataframe of hospital admissions from data warehouse}
}
\value{
copy of bundles with additional columns added giving information on adjacent admissions
}
\description{
bundle_in_spell
}

