% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/its_analysis.R
\name{create_new_vars}
\alias{create_new_vars}
\title{create_new_vars}
\usage{
create_new_vars(df = clahrcnwlhf::emergency_adms, from.vignette = FALSE,
  ward_hf_types = NULL, ward_sites = NULL, imd_lookup = NULL)
}
\arguments{
\item{df}{emergency admissions dataframe}
}
\value{
df with additional variables for ITS added
}
\description{
create_new_vars
}
