% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_analysis.R
\name{ethn_time_table}
\alias{ethn_time_table}
\title{ethn_time_table}
\usage{
ethn_time_table(df, split_by = "\%Y-\%m", ethn_col = "EthnicGroup")
}
\arguments{
\item{df}{a data frame of spells with column "EthnicGroup" or other specified
containing data on ethnic group for each spell}
\item{split_by}{date format to split by}

\item{ethn_col}{the name of the column containing the ethnicity group}
}
\value{
table of counts of ethnic groups by discharge time period
}
\description{
ethn_time_table
}
