% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_analysis.R
\name{make_emergency_adms_dataset}
\alias{make_emergency_adms_dataset}
\title{make_emergency_adms_dataset}
\usage{
make_emergency_adms_dataset(sv = TRUE, rt = FALSE)
}
\arguments{
\item{sv}{if TRUE, saves the result to the package}

\item{rt}{if TRUE, returns the results as a datframe}
}
\value{
the emergency admissions dataset with additional columns for analysis
}
\description{
make_emergency_adms_dataset
}

