% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bundle_linking.R
\name{nearest_spells}
\alias{nearest_spells}
\title{nearest_spells}
\usage{
nearest_spells(bundles, episodes)
}
\arguments{
\item{bundles}{dataframe of care bundle audit sheets}

\item{episodes}{dataframe of hospital admissions from data warehouse}
}
\value{
copy of bundles with additional columns added giving information on adjacent admissions
}
\description{
nearest_spells
}

